// src/types/mock-data.type.ts
export type MockData = {
  title: string[];
  description: string[];
  postDate: Date[];
  genre: string[];
  releaseDate: Date[];
  previewUrl: string[];
  videoUrl: string[];
  actors: string[];
  director: string[];
  userUrl: string[];
  poster: string[];
  backgroundImage: string[];
  backgroundColor: string[];
};